Unify `swift-bootstrap` with `swift-build` and `swift-test`
This is a prerequisite for enabling incremental builds when running `./Utillities/bootstrap build --release` and `./Utilities/bootstrap test --release` in sequence.
