+++ Package.swift
@@ -602,6 +602,7 @@ let package = Package(
                 .product(name: "ArgumentParser", package: "swift-argument-parser"),
                 "Basics",
                 "Build",
+                "CoreCommands",
                 "PackageGraph",
                 "PackageLoading",
                 "PackageModel",
