+++ Sources/Commands/Utilities/MultiRootSupport.swift
@@ -110,3 +110,11 @@ public struct XcodeWorkspaceLoader: WorkspaceLoader {
         }
     }
 }
+
+public extension _SwiftCommand {
+    var workspaceLoaderProvider: WorkspaceLoaderProvider {
+        return {
+            XcodeWorkspaceLoader(fileSystem: $0, observabilityScope: $1)
+        }
+    }
+}
